round 2 of cb (#2578)
