+++ tools/cb_post/benchmark.sh
@@ -17,16 +17,8 @@ echo "Download and unzip dataset"
 wget --continue https://datasets.clickhouse.com/hits_compatible/hits.json.gz
 gzip -d hits.json.gz
 
-# Add the _index line and fix the UserID from string to num and preprocesses the dataset for loading
-python3 fix_hits.py
-
-echo "Split into 10 files to increase parallelism"
-rm hits.json
-split -l 20000000 sighits.json splithits_
-rm sighits.json
-
 echo "Load data into SigLens, this can take a few hours"
-time python3 send_data.py
+time python3 send_datawithactionline.py
 
 echo "Run queries"
 ./run.sh
